requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.0.10
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
Faker==19.6.2
inflection==0.5.1
packaging==23.1
psycopg2-binary==2.9.8
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
